Move `extractNames` to its own file
